Split Luhn class
- Clear separation of concerns:
  The project is well-structured, with distinct classes for calculation (`LuhnCalculator`), validation (`LuhnValidator`), and core algorithm logic (`LuhnAlgorithm`).
  This promotes maintainability and testability.
- Extension methods:
  The use of extension methods for string manipulation and validation provides a fluent and convenient API.

Resolves: No entry
