+++ benchmark/Program.cs
@@ -1,3 +1,10 @@
-﻿using BenchmarkDotNet.Running;
+﻿namespace LuhnDotNetBenchmark;
 
-BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+using BenchmarkDotNet.Running;
+using System.Diagnostics.CodeAnalysis;
+
+[ExcludeFromCodeCoverage]
+public static class Program
+{
+    public static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+}
